[UHYU-390] refactor: 줌 레벨에 따른 검색 반경 로직을 별도 유틸로 분리 및 성능 최적화
