+++ src/shared/utils/gaTracker.ts
@@ -45,8 +45,6 @@ export const trackCustomEvent = (
 ) => {
   if (MEASUREMENT_ID) {
     ReactGA.event(eventName, parameters);
-    if (import.meta.env.MODE === 'development') {
-      console.log('[GA] Custom event tracked:', eventName, parameters);
-    }
+    console.log('[GA] Custom event tracked:', eventName, parameters);
   }
 };
